server/requirements.txt
paho-mqtt
sqlalchemy